start redesign of user database flie
